requirements.txt
pandas>=1.3
pyarrow>=6.0
pillow
pytest
